clients: Centralize Tailwind configuration in polarkit
