+++ clients/apps/chrome-extension/src/frame.css
@@ -1,3 +1,1 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+../../../packages/polarkit/src/globals.css
