+++ clients/apps/web/src/pages/_app.tsx
@@ -8,6 +8,7 @@ import Head from 'next/head'
 import { useRouter } from 'next/router'
 import { CONFIG } from 'polarkit'
 import { queryClient, QueryClientProvider } from 'polarkit/api'
+import 'polarkit/globals.css'
 import posthog from 'posthog-js'
 import { ReactElement, useEffect } from 'react'
 import '../styles/globals.scss'
