+++ clients/packages/polarkit/src/globals.css
@@ -1,0 +1,13 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@layer base {
+  html {
+    @apply text-gray-900;
+  }
+
+  html.dark {
+    @apply text-gray-200;
+  }
+}
